refactor: extract AgentSessionBar from AgentWorkspace
