+++ src/components/agent/AgentWorkspace.tsx
@@ -17,7 +17,6 @@ import { IconFinder } from "central-icons/IconFinder";
 import { IconFolder1 } from "central-icons/IconFolder1";
 import { IconFolders } from "central-icons/IconFolders";
 import { IconLightBulbSimple } from "central-icons/IconLightBulbSimple";
-import { IconConsole } from "central-icons/IconConsole";
 import { IconShieldCheck } from "central-icons/IconShieldCheck";
 import { IconStopCircle } from "central-icons/IconStopCircle";
 import { IconToolbox } from "central-icons/IconToolbox";
@@ -34,14 +33,10 @@ import { IconConsoleSimple } from "central-icons/IconConsoleSimple";
 import { IconDeepSearch } from "central-icons/IconDeepSearch";
 import { IconCheckCircle2 } from "central-icons/IconCheckCircle2";
 import { IconConcise } from "central-icons/IconConcise";
-import { IconDotGrid1x3Horizontal } from "central-icons/IconDotGrid1x3Horizontal";
 import { IconFiles } from "central-icons/IconFiles";
 import { IconFileSparkle } from "central-icons/IconFileSparkle";
 import { IconFileText } from "central-icons/IconFileText";
 import { IconEmail1Sparkle } from "central-icons/IconEmail1Sparkle";
-import { IconFolderAddRight } from "central-icons/IconFolderAddRight";
-import { IconGauge } from "central-icons/IconGauge";
-import { IconMoveFolder } from "central-icons/IconMoveFolder";
 import { IconHeartBeat } from "central-icons/IconHeartBeat";
 import { IconMagnifyingGlass } from "central-icons/IconMagnifyingGlass";
 import { IconMicrophone } from "central-icons/IconMicrophone";
@@ -51,7 +46,6 @@ import { IconPageTextSearch } from "central-icons/IconPageTextSearch";
 import { IconPencil } from "central-icons/IconPencil";
 import { IconPieChart1 } from "central-icons/IconPieChart1";
 import { IconPlusMedium } from "central-icons/IconPlusMedium";
-import { IconShareOs } from "central-icons/IconShareOs";
 import { IconShieldCrossed } from "central-icons/IconShieldCrossed";
 import { IconStop } from "central-icons/IconStop";
 import { DotSpinner } from "../DotSpinner";
@@ -73,9 +67,8 @@ import {
   useState,
   useSyncExternalStore,
 } from "react";
-import { BackButton } from "../ui/BackButton";
 import { TierMiniCard } from "../account/FundingNotice";
-import type { FundingTier, TextFundingNoticeContext } from "../account/FundingNotice";
+import type { FundingTier } from "../account/FundingNotice";
 import { ConfirmDialog } from "../ui/ConfirmDialog";
 import { CopyStateIcon } from "../ui/CopyStateIcon";
 import { Dialog } from "../ui/Dialog";
@@ -236,7 +229,6 @@ import {
 import { normalizeSteerText } from "../../lib/hermes-session-steer";
 import { buildSessionPayload } from "../../lib/share-payload";
 import { ShareDialog } from "../share/ShareDialog";
-import { ShareLinkCopyAction } from "../share/ShareLinkCopyAction";
 import { recordPositiveFeedbackSent } from "../../lib/referral-nudge";
 import { useScrollFade } from "../../lib/use-scroll-fade";
 import { unsupportedEventStore } from "../../lib/hermes-unsupported-events";
@@ -258,19 +250,13 @@ import { UnsupportedEventNotice } from "./UnsupportedEventNotice";
 import { HermesTracePanel } from "./HermesTracePanel";
 import { MarkdownContent, highlightText, type HighlightCursor } from "./MarkdownContent";
 import { SmoothedStreamingMarkdown } from "./SmoothedStreamingMarkdown";
-import {
-  ComposerModelPicker,
-  PrivacyModeBadge,
-  UnrestrictedBadge,
-  heroPrivacyFootnote,
-} from "./composer/ModelPicker";
+import { ComposerModelPicker, PrivacyModeBadge, heroPrivacyFootnote } from "./composer/ModelPicker";
 import {
   PROVIDER_MODEL_SETTINGS_CHANGED_EVENT,
   dispatchProviderModelSettingsChanged,
   modelPrivacyBadge,
   modelSupportsImageInput,
   modelSupportsTools,
-  type ModelPrivacyBadge,
   type ProviderModelSettingsChangedDetail,
 } from "../../lib/model-privacy";
 import {
@@ -383,7 +369,7 @@ import {
   rememberSessionMode,
   sessionUnrestricted,
 } from "../../lib/agent-session-modes";
-import { HERMES_TUI_DEBUG_WARNING, hermesTuiDebugAvailable } from "../../lib/hermes-tui-debug";
+import { hermesTuiDebugAvailable } from "../../lib/hermes-tui-debug";
 import {
   AGENT_CLI_ACCESS_ENABLED_MESSAGE,
   hasAgentCliAccessRequest,
@@ -415,7 +401,6 @@ import {
   prepareProjectPrompt,
   ProjectContextSignatureStore,
   stripProjectContext,
-  type AgentProjectContext,
 } from "../../lib/agent-project-context";
 import {
   buildAgentChatGallery,
@@ -423,8 +408,9 @@ import {
   type AgentChatGallerySection,
 } from "../../lib/agent-chat-gallery";
 import { attachScrollThumbFade } from "../../lib/scroll-thumb-fade";
-import type { AgentWorkspaceOrigin, AgentWorkspaceProps } from "./agent-workspace-types";
+import type { AgentWorkspaceProps } from "./agent-workspace-types";
 export type { AgentWorkspaceOrigin } from "./agent-workspace-types";
+import { AgentSessionBar } from "./chat-turns/AgentSessionBar";
 import {
   upstreamProviderRecoveryIds,
   upstreamProviderRecoveryStore,
@@ -12295,332 +12281,6 @@ export function AgentWorkspace({
     </section>
   );
 }
-
-// Persistent, full-width session bar — same chrome as the Notes/Folders
-// breadcrumb. Stays pinned while the conversation scrolls beneath it, carries
-// the back arrow + origin crumbs (Projects / {project} or Agents), the
-// private-mode badge, and folds rename/delete into an overflow menu so the
-// conversation keeps the focus (no separate title heading).
-function AgentSessionBar({
-  origin,
-  privacyBadge,
-  fullMode,
-  title,
-  shareUrl,
-  artifactCount = 0,
-  artifactsOpen = false,
-  inProject = false,
-  projectContext,
-  onToggleArtifacts,
-  onRename,
-  onShare,
-  onMoveToProject,
-  onDelete,
-  onShowUsage,
-  onCompactContext,
-  onOpenTuiDebug,
-}: {
-  origin?: AgentWorkspaceOrigin;
-  privacyBadge?: ModelPrivacyBadge;
-  fullMode?: boolean;
-  title?: string;
-  shareUrl?: string;
-  artifactCount?: number;
-  artifactsOpen?: boolean;
-  inProject?: boolean;
-  projectContext?: AgentProjectContext;
-  onToggleArtifacts?: () => void;
-  onRename?: (title: string) => void;
-  /** Opens the private-sharing dialog for this session (JUN-308). */
-  onShare?: () => void;
-  /** Opens the change-project dialog (which also owns removal). */
-  onMoveToProject?: () => void;
-  onDelete?: () => void;
-  onShowUsage?: () => void;
-  onCompactContext?: () => void;
-  /** Developer-only: open this session in Hermes' raw TUI. Undefined (and the
-   * menu item absent) in production builds. */
-  onOpenTuiDebug?: () => void;
-}) {
-  const [renaming, setRenaming] = useState(false);
-  const [draft, setDraft] = useState(title ?? "");
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [instructionsOpen, setInstructionsOpen] = useState(false);
-  const menuWrapRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    function onPointer(event: MouseEvent) {
-      if (!menuWrapRef.current?.contains(event.target as Node)) {
-        setMenuOpen(false);
-      }
-    }
-    function onKey(event: KeyboardEvent) {
-      if (event.key === "Escape") setMenuOpen(false);
-    }
-    window.addEventListener("mousedown", onPointer);
-    window.addEventListener("keydown", onKey);
-    return () => {
-      window.removeEventListener("mousedown", onPointer);
-      window.removeEventListener("keydown", onKey);
-    };
-  }, [menuOpen]);
-
-  function commitRename() {
-    setRenaming(false);
-    onRename?.(draft);
-  }
-
-  const hasMenu = Boolean(
-    onRename ||
-      onShare ||
-      onMoveToProject ||
-      onDelete ||
-      onShowUsage ||
-      onCompactContext ||
-      onOpenTuiDebug,
-  );
-
-  return (
-    <div className="detail-bar agent-session-bar" data-tauri-drag-region>
-      {origin ? <BackButton label={origin.backLabel} onClick={origin.onBack} /> : null}
-      <nav className="detail-breadcrumb" aria-label="Breadcrumb">
-        <ol>
-          {origin ? (
-            origin.crumbs.map((crumb, index) => (
-              <li key={`${crumb.label}-${index}`}>
-                {index > 0 ? (
-                  <span className="detail-breadcrumb-separator" aria-hidden>
-                    /
-                  </span>
-                ) : null}
-                <button type="button" className="detail-breadcrumb-link" onClick={crumb.onClick}>
-                  {crumb.icon ? (
-                    <span className="detail-breadcrumb-icon" aria-hidden>
-                      {crumb.icon}
-                    </span>
-                  ) : null}
-                  {crumb.label}
-                </button>
-              </li>
-            ))
-          ) : (
-            <li>
-              <span className="detail-breadcrumb-label">Session</span>
-            </li>
-          )}
-          {title !== undefined ? (
-            <li>
-              <span className="detail-breadcrumb-separator" aria-hidden>
-                /
-              </span>
-              {renaming ? (
-                <input
-                  className="agent-session-rename"
-                  aria-label="Session name"
-                  autoFocus
-                  value={draft}
-                  onChange={(event) => setDraft(event.currentTarget.value)}
-                  onBlur={commitRename}
-                  onKeyDown={(event) => {
-                    if (event.key === "Enter") {
-                      event.preventDefault();
-                      commitRename();
-                    }
-                    if (event.key === "Escape") {
-                      setRenaming(false);
-                      setDraft(title ?? "");
-                    }
-                  }}
-                />
-              ) : (
-                <span className="detail-breadcrumb-current-group">
-                  <span className="detail-breadcrumb-current">{title || "Untitled session"}</span>
-                  {shareUrl ? <ShareLinkCopyAction url={shareUrl} /> : null}
-                </span>
-              )}
-            </li>
-          ) : origin ? (
-            <li>
-              <span className="detail-breadcrumb-separator" aria-hidden>
-                /
-              </span>
-              <span className="detail-breadcrumb-current">New session</span>
-            </li>
-          ) : null}
-        </ol>
-      </nav>
-      <div className="detail-bar-actions">
-        {projectContext ? (
-          <button
-            type="button"
-            className="agent-project-instructions"
-            onClick={() => setInstructionsOpen(true)}
-          >
-            Project instructions
-          </button>
-        ) : null}
-        {fullMode ? <UnrestrictedBadge /> : null}
-        {onToggleArtifacts && artifactCount > 0 ? (
-          <button
-            type="button"
-            className="agent-session-files"
-            aria-label={`View files (${artifactCount})`}
-            title="View files"
-            aria-pressed={artifactsOpen}
-            onClick={onToggleArtifacts}
-          >
-            <IconFiles size={14} />
-            <span aria-hidden>{artifactCount}</span>
-          </button>
-        ) : null}
-        <PrivacyModeBadge badge={privacyBadge} />
-        {hasMenu ? (
-          <div className="agent-session-menu-wrap" ref={menuWrapRef}>
-            <button
-              type="button"
-              className="icon-button agent-session-menu-trigger"
-              aria-label="Session actions"
-              aria-haspopup="menu"
-              aria-expanded={menuOpen}
-              onClick={() => setMenuOpen((open) => !open)}
-            >
-              <IconDotGrid1x3Horizontal size={16} />
-            </button>
-            {menuOpen ? (
-              <div className="sidebar-identity-menu agent-session-menu" role="menu">
-                {onRename ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      setDraft(title ?? "");
-                      setRenaming(true);
-                    }}
-                  >
-                    <IconPencil size={14} />
-                    Rename
-                  </button>
-                ) : null}
-                {onShare ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onShare();
-                    }}
-                  >
-                    <IconShareOs size={14} />
-                    Share
-                  </button>
-                ) : null}
-                {onMoveToProject ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onMoveToProject();
-                    }}
-                  >
-                    {inProject ? <IconMoveFolder size={14} /> : <IconFolderAddRight size={14} />}
-                    {inProject ? "Change project" : "Add to project"}
-                  </button>
-                ) : null}
-                {(onRename || onShare || onMoveToProject) && (onShowUsage || onCompactContext) ? (
-                  <div className="context-menu-separator" role="separator" />
-                ) : null}
-                {onShowUsage ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onShowUsage();
-                    }}
-                  >
-                    <IconGauge size={14} />
-                    Usage
-                  </button>
-                ) : null}
-                {onCompactContext ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onCompactContext();
-                    }}
-                  >
-                    <IconConcise size={14} />
-                    Compact context
-                  </button>
-                ) : null}
-                {onDelete && (onRename || onMoveToProject || onShowUsage || onCompactContext) ? (
-                  <div className="context-menu-separator" role="separator" />
-                ) : null}
-                {onDelete ? (
-                  <button
-                    type="button"
-                    role="menuitem"
-                    className="destructive"
-                    onClick={() => {
-                      setMenuOpen(false);
-                      onDelete();
-                    }}
-                  >
-                    <IconTrashCan size={14} />
-                    Delete session
-                  </button>
-                ) : null}
-                {onOpenTuiDebug ? (
-                  <>
-                    <div className="context-menu-separator" role="separator" />
-                    <button
-                      type="button"
-                      role="menuitem"
-                      // Debug-only fallback: resume this session in Hermes' raw
-                      // TUI to tell a June adapter/UI bug from a Hermes one.
-                      title={HERMES_TUI_DEBUG_WARNING}
-                      onClick={() => {
-                        setMenuOpen(false);
-                        onOpenTuiDebug();
-                      }}
-                    >
-                      <IconConsole size={14} />
-                      Debug with Hermes TUI
-                    </button>
-                  </>
-                ) : null}
-              </div>
-            ) : null}
-          </div>
-        ) : null}
-      </div>
-      <Dialog
-        open={instructionsOpen}
-        onClose={() => setInstructionsOpen(false)}
-        title={`${projectContext?.name ?? "Project"} instructions`}
-        footer={
-          <button
-            type="button"
-            className="primary-action"
-            onClick={() => setInstructionsOpen(false)}
-          >
-            Close
-          </button>
-        }
-      >
-        <div className="agent-project-instructions-content">
-          {projectContext?.instructions?.trim() || "No project instructions have been added."}
-        </div>
-      </Dialog>
-    </div>
-  );
-}
-
 const AGENT_TITLE_MAX_CHARS = 48;
 
 async function agentSessionTitleForPrompt(prompt: string, response?: string) {
